Add request information to NetworkError
APIErrors do not have this issue because you get the response
information and that contains the request information. However
NetworkError only contained a error message and had no information on
where it came from or what it related to. By passing in the request
parameters you are sending to the fetch request it makes it easy to do
say a network error occured connecting to x url etc. Also allow someone
to check and work with headers if there were any issues. Generally makes
debugging easier.