refactor: Use Promises instead of async/await
Using async/ await introduces a dependency on regenerator-runtime, which is discouraged on third-party libraries. Since @hubble/request is effectively becoming a third-party library, we should switch it to use Promise in its raw form, rather than async/await.

@hubble/request will still be fully compatible with async/await, but it won’t use it itself. (I’ve left the unit tests using async/await to demonstrate this.)

It’s difficult to use regenerator-runtime as a library. It’s provided by Babel, but packages that are published on NPM aren’t normally compiled by Babel. Additionally, regenerator-runtime is only provided by Next.js in the browser - they explicitly disable it on the server