Downgrade `isomorphic-unfetch` package from 4.0.2 to 4.0.0
v4.0.1 introduced a "Module not found: Error: Can't resolve 'unfetch' in '/src/node_modules/isomorphic-unfetch'" error when trying to compile projects using our updated @hubble/request package. This was due to it using v5.0.0 of the `unfetch` package as a dependency, which has an invalid path export in its package.json.
There's a PR for a fix submitted, but as it's been sitting there since 10/02/2023, there's no telling how long this may take to be updated. Therefore, downgrading to v4.0.0 of `isomorphic-unfetch` ensures it uses v4.2.0 of `unfetch`, which doesn't seem to have this problem.