Change the target in the cloudbuild.yaml to dev
There were failures because it relied on eslint being installed which is a dev dependency. The main reason we run checks against a production container is to ensure that the exact deployed environment works. Here, we only need build-and-test step as deployment is a manual process so the build isn't producing a production container (I THINK!) and so it should be ok to run tests and linting against a dev env.